+++ src/common/util/getSessionFetch.ts
@@ -19,10 +19,29 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import type { VerifiableCredential } from "@inrupt/solid-client-vc";
-import type { AccessGrantBody } from "./AccessVerifiableCredential";
+import { fetch as crossFetch } from "cross-fetch";
+import { AccessBaseOptions } from "../../gConsent/type/AccessBaseOptions";
 
-export type AccessGrant = VerifiableCredential & AccessGrantBody;
-// Alias with the explicit GConsent suffix, to be used in functions accepting both
-// gConsent-based and ODRL-based grants.
-export type AccessGrantGConsent = AccessGrant;
+/**
+ * Dynamically import solid-client-authn-browser so that
+ * this library doesn't have a hard dependency.
+ *
+ * @returns fetch function
+ */
+export async function getSessionFetch(
+  options: AccessBaseOptions
+): Promise<typeof fetch> {
+  if (options.fetch) {
+    return options.fetch;
+  }
+  try {
+    const { fetch: fetchFn } = await import(
+      "@inrupt/solid-client-authn-browser"
+    );
+
+    return fetchFn;
+  } catch (e) {
+    /* istanbul ignore next: @inrupt/solid-client-authn-browser is a devDependency, so this path is not hit in tests: */
+    return crossFetch;
+  }
+}