+++ src/common/verify/isValidAccessGrant.ts
@@ -25,7 +25,6 @@ import {
   VerifiableCredential,
   getVerifiableCredentialApiConfiguration,
 } from "@inrupt/solid-client-vc";
-import type { AccessBaseOptions } from "../type/AccessBaseOptions";
 import { getSessionFetch } from "../util/getSessionFetch";
 
 /**
@@ -39,8 +38,9 @@ import { getSessionFetch } from "../util/getSessionFetch";
 // TODO: Push verification further as this just checks it's a valid VC should we not type check the consent grant?
 async function isValidAccessGrant(
   vc: VerifiableCredential | URL | UrlString,
-  options: Exclude<AccessBaseOptions, "accessEndpoint"> & {
+  options: {
     verificationEndpoint?: UrlString;
+    fetch?: typeof fetch;
   } = {}
 ): Promise<{ checks: string[]; warnings: string[]; errors: string[] }> {
   const fetcher = await getSessionFetch(options);