+++ src/odrl/index.ts
@@ -19,5 +19,11 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-// TODO Remove when actually exporting functions from the odrl module.
-export const DUMMY_CONSTANT = 0;
+export type {
+  AccessGrantOdrl,
+  CredentialSubjectOdrl,
+  OdrlConstraint,
+  OdrlPermission,
+} from "./type/AccessGrant";
+
+export { isCredentialAccessGrantOdrl } from "./type/AccessGrant";